# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Browser Automation Service API",
    "description": "A service for automating browser actions through two execution modes: Extension-based (WebSocket connected browser extension) and Server-side (headless Chromium). The service automatically chooses the appropriate mode based on connection availability.",
    "version": "1.0.0",
    "contact": {
      "name": "Browser Automation Service",
      "url": "http://10.0.0.2:3010"
    }
  },
  "servers": [
    {
      "url": "http://10.0.0.2:3010",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3010",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/sessions": {
      "get": {
        "tags": ["Sessions"],
        "summary": "List all active sessions",
        "description": "Returns a list of all currently active browser automation sessions",
        "operationId": "getSessions",
        "responses": {
          "200": {
            "description": "List of active sessions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sessions": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Session"
                      }
                    },
                    "count": {
                      "type": "integer",
                      "description": "Total number of active sessions"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Sessions"],
        "summary": "Create a new session",
        "description": "Creates a new browser automation session",
        "operationId": "createSession",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              },
              "examples": {
                "basic": {
                  "summary": "Basic session creation",
                  "value": {
                    "metadata": {
                      "browser": "chrome",
                      "purpose": "testing"
                    }
                  }
                },
                "detailed": {
                  "summary": "Detailed session with options",
                  "value": {
                    "metadata": {
                      "browser": "chrome",
                      "purpose": "automated-testing",
                      "user": "test-team",
                      "project": "web-automation"
                    },
                    "options": {
                      "timeout": 60000,
                      "maxCommands": 100
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Session created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Too many sessions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{sessionId}": {
      "get": {
        "tags": ["Sessions"],
        "summary": "Get session details",
        "description": "Returns details about a specific session",
        "operationId": "getSession",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unique session identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Session details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "404": {
            "description": "Session not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Sessions"],
        "summary": "Delete a session",
        "description": "Terminates and removes a browser automation session",
        "operationId": "deleteSession",
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Unique session identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Session deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Session deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Session not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/extension/download": {
      "get": {
        "tags": ["Extension"],
        "summary": "Download browser extension",
        "description": "Downloads the browser extension package as a ZIP file",
        "operationId": "downloadExtension",
        "responses": {
          "200": {
            "description": "Extension package",
            "content": {
              "application/zip": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            },
            "headers": {
              "Content-Disposition": {
                "schema": {
                  "type": "string"
                },
                "description": "attachment; filename=browser-automation-extension.zip"
              }
            }
          },
          "404": {
            "description": "Extension package not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Session": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique session identifier"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Session creation timestamp"
          },
          "lastActivity": {
            "type": "string",
            "format": "date-time",
            "description": "Last activity timestamp"
          },
          "metadata": {
            "type": "object",
            "description": "Session metadata provided during creation"
          },
          "status": {
            "type": "string",
            "enum": ["active", "idle", "error"],
            "description": "Current session status"
          },
          "commandCount": {
            "type": "integer",
            "description": "Number of commands executed in this session"
          }
        },
        "required": ["id", "createdAt", "status"]
      },
      "CreateSessionRequest": {
        "type": "object",
        "properties": {
          "metadata": {
            "type": "object",
            "properties": {
              "browser": {
                "type": "string",
                "example": "chrome",
                "description": "Browser type"
              },
              "purpose": {
                "type": "string",
                "example": "testing",
                "description": "Purpose of the session"
              },
              "user": {
                "type": "string",
                "description": "User or team identifier"
              },
              "project": {
                "type": "string",
                "description": "Project identifier"
              }
            },
            "required": ["browser", "purpose"]
          },
          "options": {
            "type": "object",
            "properties": {
              "timeout": {
                "type": "integer",
                "default": 30000,
                "description": "Session timeout in milliseconds"
              },
              "maxCommands": {
                "type": "integer",
                "default": 100,
                "description": "Maximum number of commands per session"
              }
            }
          }
        },
        "required": ["metadata"]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "code": {
            "type": "string",
            "description": "Error code"
          },
          "details": {
            "type": "object",
            "description": "Additional error details"
          }
        },
        "required": ["error"]
      }
    }
  }
}